feat(react-blessed): intialize
still need to customize the view and transfer all the logic to work within the application.
