+++ .babelrc
@@ -1,0 +1,8 @@
+{
+    "presets": [
+        "env",
+        "es2015",
+        "stage-0",
+        "react"
+    ]
+}
