+++ package.json
@@ -4,23 +4,33 @@
   "description": "coinbase recommendation bot",
   "main": "src/index.js",
   "scripts": {
-    "go": "node index.js",
-    "dev": "nodemon index.js",
-    "test": "jest"
+    "go": "node --exec npm run babel-node -- index.js",
+    "dev": "nodemon --exec npm run babel-node -- index.js",
+    "test": "jest",
+    "babel-node": "babel-node"
   },
   "jest": {
     "testEnvironment": "node"
   },
   "dependencies": {
+    "blessed": "^0.1.81",
     "chalk": "^1.1.3",
     "dotenv": "^4.0.0",
     "flow": "^0.2.3",
     "gdax": "coinbase/gdax-node",
     "global": "^4.3.1",
     "moment": "^2.18.1",
+    "react": "^0.14.7",
+    "react-blessed": "^0.1.8",
     "twilio": "^2.11.1"
   },
   "devDependencies": {
+    "babel-cli": "^6.24.1",
+    "babel-preset-env": "^1.4.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-0": "^6.24.1",
+    "babel-register": "^6.24.1",
     "flow-bin": "^0.43.1",
     "flow-remove-types": "^1.2.0",
     "jest": "^19.0.2",
